requirements.txt
sqlalchemy 
psycopg2-binary
flask
python-dotenv
fastapi 
uvicorn
pydantic

